# fix(DicomMetadataStore): improved performance