--- a/.docker/Nginx-Orthanc/volumes/orthanc-db/.gitignore
+++ b/.docker/Nginx-Orthanc/volumes/orthanc-db/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore