--- a/extensions/cornerstone/src/services/ToolGroupService/ToolGroupService.ts
+++ b/extensions/cornerstone/src/services/ToolGroupService/ToolGroupService.ts
@@ -290,11 +290,7 @@ export default class ToolGroupService {
     }
 
     const filteredViewports = Array.from(viewports.values()).filter(viewport => {
-      if (!viewport.viewportOptions) {
-        return false;
-      }
-
-      return toolGroupViewportIds.includes(viewport.viewportOptions.viewportId);
+      return toolGroupViewportIds.includes(viewport.viewportId);
     });
 
     if (!filteredViewports.length) {