--- a/extensions/cornerstone/src/utils/measurementServiceMappings/Bidirectional.ts
+++ b/extensions/cornerstone/src/utils/measurementServiceMappings/Bidirectional.ts
@@ -99,8 +99,7 @@ function getMappedAnnotations(annotation, displaySetService) {
     );
 
     const { SeriesNumber } = displaySet;
-    const { length, width } = targetStats;
-    const unit = 'mm';
+    const { length, width, unit } = targetStats;
 
     annotations.push({
       SeriesInstanceUID,
@@ -130,9 +129,9 @@ function _getReport(mappedAnnotations, points, FrameOfReferenceUID) {
   values.push('Cornerstone:Bidirectional');
 
   mappedAnnotations.forEach(annotation => {
-    const { length, width } = annotation;
-    columns.push(`Length (mm)`, `Width (mm)`);
-    values.push(length, width);
+    const { length, width, unit } = annotation;
+    columns.push(`Length`, `Width`, 'Unit');
+    values.push(length, width, unit);
   });
 
   if (FrameOfReferenceUID) {
@@ -162,7 +161,14 @@ function getDisplayText(mappedAnnotations, displaySet) {
   const displayText = [];
 
   // Area is the same for all series
-  const { length, width, SeriesNumber, SOPInstanceUID, frameNumber } = mappedAnnotations[0];
+  const {
+    length,
+    width,
+    unit,
+    SeriesNumber,
+    SOPInstanceUID,
+    frameNumber,
+  } = mappedAnnotations[0];
   const roundedLength = utils.roundNumber(length, 2);
   const roundedWidth = utils.roundNumber(width, 2);
 
@@ -176,8 +182,10 @@ function getDisplayText(mappedAnnotations, displaySet) {
   const instanceText = InstanceNumber ? ` I: ${InstanceNumber}` : '';
   const frameText = displaySet.isMultiFrame ? ` F: ${frameNumber}` : '';
 
-  displayText.push(`L: ${roundedLength} mm (S: ${SeriesNumber}${instanceText}${frameText})`);
-  displayText.push(`W: ${roundedWidth} mm`);
+  displayText.push(
+    `L: ${roundedLength} ${unit} (S: ${SeriesNumber}${instanceText}${frameText})`
+  );
+  displayText.push(`W: ${roundedWidth} ${unit}`);
 
   return displayText;
 }