--- a/lerna.json
+++ b/lerna.json
@@ -1,5 +1,5 @@
 {
-  "version": "3.7.0-beta.73",
+  "version": "3.7.0-beta.76",
   "packages": ["extensions/*", "platform/*", "modes/*"],
   "npmClient": "yarn"
 }