--- a/platform/app/.recipes/OpenResty-Orthanc-Keycloak/config/nginx.conf
+++ b/platform/app/.recipes/OpenResty-Orthanc-Keycloak/config/nginx.conf
@@ -22,7 +22,7 @@ http {
 
     # lua_ settings
     #
-    lua_package_path '/usr/local/openresty/lualib/?.lua;;';
+    lua_package_path '/usr/local/openresty/lualib/?.lua;;/usr/local/share/lua/5.4/?.lua;;';
     lua_shared_dict discovery 1m; # cache for discovery metadata documents
     lua_shared_dict jwks 1m;      # cache for JWKs
     # lua_ssl_trusted_certificate /etc/ssl/certs/ca-certificates.crt;
@@ -182,6 +182,8 @@ http {
           index index.html;
           try_files $uri $uri/ /index.html;
           add_header Cache-Control "no-store, no-cache, must-revalidate";
+          add_header 'Cross-Origin-Opener-Policy' 'same-origin' always;
+      	  add_header 'Cross-Origin-Embedder-Policy' 'require-corp' always;
         }
 
         # EXAMPLE: Reverse Proxy, no auth