Some methods were taking a lot of time to run due to a O(n) search (instances).
These are the times for a study with 9400 instances:
  - getTimePointsRequests: +12s > 83ms
  - get4DVolumeInfo: +9s > 79ms
  - addInstances : +3s > 245ms